Move Google API code to separate base fragment
